Systems for woodcutting mostly in place. Going to do  the rest of the major debugging tomorrow. As is once picking up their axe's entities don't move toward trees

+++ Dwarf/ECS/Systems/Gui/InputHandler.cpp
@@ -201,7 +201,7 @@ void InputHandler::designate(const int key)
 		{
 			int state = guiState.itemSelected;
 
-			emit(designation_message{ std::make_pair(designateIdx, clickIdx), state });
+			emit(designation_message{ std::make_pair(designateIdx, clickIdx), static_cast<uint8_t>(state) });
 			designateIdx = 0;
 		}
 	}
